// client/src/features/TPI/types/asset.types.ts
export interface TPIAsset {
  id: string;
  name: string;
  cmdbStatus: string;
  assetType: string;
  affinityGroup: string;
  appApprModernDelivery: string;
  applicationTypeFinancial: string;
  architect: string;
  assetIdInFAST: string;
  assetIdInSchedule: string;
  assetIdInWeeklyStatusReport: string;
  assetTier: string;
  blockFunding: string;
  btoAlignment: string;
  businessOwnerCommsCheck: string;
  businessOwnerOwnedBy: string;
  businessOwnerSME: string;
  cashPaymentSystems: string;
  cmdbBeingRetired: string;
  cmdbLegalHold: string;
  concatinatedBTOandDivision: string;
  connectorStatus: string;
  cotsOrInHouseBuilt: string;
  customerFacing: string;
  defaultTier: string;
  description: string;
  disposition: string;
  externalFacing: string;
  financialImpact4hrOutage: string;
  foundational: string;
  highLevelBTO: string;
  hosted: string;
  infoSecCritical: string;
  informationClassification: string;
  isSaas: string;
  itOwnerCommsCheck: string;
  itOwnerManagedBy: string;
  keyChainOnboardingStatus: string;
  maintenanceWindow: string;
  mdAssetDesignation: string;
  multiFactorAuthentication: string;
  nfr9: string;
  nfr10: string;
  nonDefaultTier1: string;
  nonDefaultTier2: string;
  nonDefaultTier3: string;
  nonDefaultTier4: string;
  onboardingStatus: string;
  operationalHours: string;
  owningInternalOrg: string;
  ppiClassification: string;
  privilegedAccess: string;
  sox: string;
  spof: string;
  sppi: string;
  status: string;
  supportSME: string;
  supportedBy: string;
  supportedByCommsCheck: string;
  version: string;
  createdAt?: string;
}

export interface TPIHistoryRecord {
  id: number;
  tpiAssetId: string;
  startDate: string;
  endDate: string;
  name: string;
  cmdbStatus: string;
  assetType: string;
  affinityGroup: string;
  appApprModernDelivery: string;
  applicationTypeFinancial: string;
  architect: string;
  assetIdInFAST: string;
  assetIdInSchedule: string;
  assetIdInWeeklyStatusReport: string;
  assetTier: string;
  blockFunding: string;
  btoAlignment: string;
  businessOwnerCommsCheck: string;
  businessOwnerOwnedBy: string;
  businessOwnerSME: string;
  cashPaymentSystems: string;
  cmdbBeingRetired: string;
  cmdbLegalHold: string;
  concatinatedBTOandDivision: string;
  connectorStatus: string;
  cotsOrInHouseBuilt: string;
  customerFacing: string;
  defaultTier: string;
  description: string;
  disposition: string;
  externalFacing: string;
  financialImpact4hrOutage: string;
  foundational: string;
  highLevelBTO: string;
  hosted: string;
  infoSecCritical: string;
  informationClassification: string;
  isSaas: string;
  itOwnerCommsCheck: string;
  itOwnerManagedBy: string;
  keyChainOnboardingStatus: string;
  maintenanceWindow: string;
  mdAssetDesignation: string;
  multiFactorAuthentication: string;
  nfr9: string;
  nfr10: string;
  nonDefaultTier1: string;
  nonDefaultTier2: string;
  nonDefaultTier3: string;
  nonDefaultTier4: string;
  onboardingStatus: string;
  operationalHours: string;
  owningInternalOrg: string;
  ppiClassification: string;
  privilegedAccess: string;
  sox: string;
  spof: string;
  sppi: string;
  status: string;
  supportSME: string;
  supportedBy: string;
  supportedByCommsCheck: string;
  version: string;
}

export interface TPIHistoryResponse {
  history: TPIHistoryRecord[];
  total: number;
}